refactor(recipes): move builders out of train_ft.py and vlm/finetune.py
Moves ~615 lines from recipes/llm/train_ft.py and ~340 lines from
recipes/vlm/finetune.py into components/:

- components/training/build.py: build_model (LLM), build_vlm_model
- components/datasets/llm/build.py: build_dataloader, build_validation_dataloader,
  _uses_te_dot_product_attention, _uses_thd_collater, _get_num_thd_chunks
- components/datasets/vlm/build.py: build_vlm_dataloader
- components/loss/calculate.py: calculate_loss
- components/loss/mtp.py: calculate_mtp_loss
- _transformers/auto_tokenizer.py: _build_tokenizer, _get_model_name,
  compute_trust_remote_code_from_model

Recipes keep back-compat re-imports so kd.py and external test patches
continue to resolve. Test patch sites for build_dataloader and
resolve_trust_remote_code updated to point at the new module locations.
