+++ nemo_automodel/_transformers/auto_tokenizer.py
@@ -134,10 +134,96 @@ class NeMoAutoTokenizer:
         )
 
 
+# ── Recipe helpers (model-name / trust_remote_code / tokenizer-from-cfg) ──
+# These used to live in recipes/llm/train_ft.py. They're tokenizer/model-config
+# utilities used by the recipe and by external callers; they belong next to
+# NeMoAutoTokenizer.
+
+
+def _get_model_name(cfg_model):
+    """Pull ``pretrained_model_name_or_path`` from a recipe-style model config.
+
+    Accepts either a top-level ``pretrained_model_name_or_path`` field, or a
+    nested ``config`` (string or sub-config) carrying the same field. Returns
+    ``None`` if absent.
+    """
+    if cfg_model.get("pretrained_model_name_or_path", None) is not None:
+        return cfg_model.pretrained_model_name_or_path
+    elif cfg_model.get("config", None) is not None:
+        if isinstance(cfg_model.config, str):
+            return cfg_model.config
+        return cfg_model.config.get("pretrained_model_name_or_path", None)
+    else:
+        return None
+
+
+def compute_trust_remote_code_from_model(cfg_model):
+    """Compute ``trust_remote_code`` from a recipe-style model config.
+
+    Reads ``cfg_model.trust_remote_code`` or ``cfg_model.config.trust_remote_code``
+    if present; otherwise falls back to
+    :func:`nemo_automodel.components.utils.model_utils.resolve_trust_remote_code`
+    against the model name.
+    """
+    from nemo_automodel.components.utils.model_utils import resolve_trust_remote_code
+
+    if hasattr(cfg_model, "trust_remote_code"):
+        return getattr(cfg_model, "trust_remote_code")
+    elif hasattr(cfg_model, "config") and hasattr(cfg_model.config, "trust_remote_code"):
+        return getattr(cfg_model.config, "trust_remote_code")
+    return resolve_trust_remote_code(_get_model_name(cfg_model))
+
+
+def _build_tokenizer(cfg_model, cfg_ds):
+    """Build a tokenizer from the recipe's model + dataset config.
+
+    Resolution order:
+      1. If ``cfg_ds.tokenizer`` is absent and ``cfg_model`` provides a model name,
+         load via :class:`NeMoAutoTokenizer.from_pretrained`.
+      2. Else if ``cfg_ds.tokenizer`` is a plain config (no ``_target_``), forward
+         its kwargs to ``NeMoAutoTokenizer.from_pretrained``.
+      3. Else if ``cfg_ds.tokenizer`` has ``_target_``, instantiate the configured
+         class.
+
+    Returns ``(kwargs_for_dataset, tokenizer)`` where ``kwargs_for_dataset`` is
+    a dict ready to merge into the dataset constructor — when the dataset's
+    target accepts a ``tokenizer`` argument, the tokenizer is included.
+    """
+    import inspect as _inspect
+
+    trust_remote_code = compute_trust_remote_code_from_model(cfg_model)
+    if "tokenizer" not in cfg_ds and _get_model_name(cfg_model) is not None:
+        logger.info("Using model config to instantiate tokenizer")
+        tokenizer = NeMoAutoTokenizer.from_pretrained(_get_model_name(cfg_model), trust_remote_code=trust_remote_code)
+    elif cfg_ds.get("tokenizer", None) is None:
+        tokenizer = None
+    elif "_target_" not in cfg_ds.tokenizer:
+        tokenizer_dict = cfg_ds.tokenizer.to_dict()
+        trust_remote_code = tokenizer_dict.pop("trust_remote_code", trust_remote_code)
+        tokenizer = NeMoAutoTokenizer.from_pretrained(**tokenizer_dict, trust_remote_code=trust_remote_code)
+    else:
+        trust_remote_code = cfg_ds.tokenizer.to_dict().pop("trust_remote_code", trust_remote_code)
+        tokenizer = cfg_ds.tokenizer.instantiate(trust_remote_code=trust_remote_code)
+
+    # Pass tokenizer into the dataset constructor only if it accepts it.
+    kwargs = {}
+    if tokenizer is not None and callable(cfg_ds._target_):
+        try:
+            sig = _inspect.signature(cfg_ds._target_)
+            if "tokenizer" in sig.parameters:
+                kwargs["tokenizer"] = tokenizer
+        except (ValueError, TypeError):
+            pass
+    return kwargs, tokenizer
+
+
 __all__ = [
     "NeMoAutoTokenizer",
     "NeMoAutoTokenizerWithBosEosEnforced",
     "TokenizerRegistry",
+    "_build_tokenizer",
+    "_get_model_name",
+    "compute_trust_remote_code_from_model",
 ]
