+++ nemo_automodel/recipes/llm/train_ft.py
@@ -96,628 +96,26 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-# ---------------------------
-#  Stateless helper functions
-# ---------------------------
-def _get_model_name(cfg_model):
-    if cfg_model.get("pretrained_model_name_or_path", None) is not None:
-        return cfg_model.pretrained_model_name_or_path
-    elif cfg_model.get("config", None) is not None:
-        if isinstance(cfg_model.config, str):
-            return cfg_model.config
-        return cfg_model.config.get("pretrained_model_name_or_path", None)
-    else:
-        return None
-
-
-def _uses_te_dot_product_attention(model_or_cfg):
-    """Check whether the model uses TE DotProductAttention.
-
-    Accepts either an instantiated nn.Module (preferred — inspects actual modules)
-    or a config object (fallback — checks backend.attn string).
-    """
-    if isinstance(model_or_cfg, torch.nn.Module):
-        try:
-            from transformer_engine.pytorch.attention import DotProductAttention
-        except ImportError:
-            return False
-        return any(isinstance(m, DotProductAttention) for m in model_or_cfg.modules())
-    # Config fallback for call sites before model is built
-    return (
-        hasattr(model_or_cfg, "backend") and hasattr(model_or_cfg.backend, "attn") and model_or_cfg.backend.attn == "te"
-    )
-
-
-def _uses_thd_collater(cfg_dataloader):
-    from nemo_automodel.components.datasets.utils import packed_sequence_thd_collater
-
-    return (
-        True
-        if hasattr(cfg_dataloader, "collate_fn") and cfg_dataloader.collate_fn == packed_sequence_thd_collater
-        else False
-    )
-
-
-def _get_num_thd_chunks(pp_enabled, cfg):
-    if pp_enabled:
-        return cfg.step_scheduler.local_batch_size // cfg.get("distributed.pipeline.pp_microbatch_size", 1)
-    return 1
-
-
-def build_model(
-    cfg_model,
-    cfg_peft,
-    seed,
-    has_packed_sequence=False,
-    cfg_fp8=None,
-    cfg_compile=None,
-    cfg_quantization=None,
-    device_mesh=None,
-    moe_mesh=None,
-    distributed_config=None,
-    pipeline_config=None,
-    cfg_qat=None,
-    cfg_moe=None,
-    activation_checkpointing=False,
-    unfreeze_modules: list[str] | None = None,
-    sdpa_method: list[str] | None = None,
-) -> tuple[nn.Module | AutoPipeline, list["Optimizer"]]:  # noqa: F821
-    """Build and initialize a model.
-
-    Args:
-        cfg_model: Configuration for model instantiation.
-        cfg_peft: Configuration for PEFT.
-        seed: Random seed.
-        has_packed_sequence: Whether using packed sequences.
-        cfg_fp8: Configuration for FP8.
-        cfg_compile: Configuration for torch.compile.
-        cfg_quantization: Configuration for BitsAndBytes quantization.
-        device_mesh: Device mesh for distributed training.
-        moe_mesh: MOE mesh for expert parallelism.
-        distributed_config: Strategy-specific distributed config (FSDP2Config, etc.).
-        pipeline_config: Pipeline parallelism config.
-        cfg_qat: Configuration for QAT (will be instantiated to QATConfig).
-        cfg_moe: MoEParallelizerConfig instance, or ConfigNode to be converted.
-        activation_checkpointing: Whether to enable activation checkpointing.
-        unfreeze_modules: List of module names/substrings to unfreeze.
-        sdpa_method: Explicit list of SDPA backend name strings (e.g.
-            ``["flash_attention", "efficient_attention"]``), or ``None`` to
-            auto-select based on CP / activation checkpointing.
-    """
-    with ScopedRNG(seed=seed, ranked=True):
-        kwargs = {
-            "has_packed_sequence": has_packed_sequence,
-            "peft_config": cfg_peft,
-            "device_mesh": device_mesh,
-            "moe_mesh": moe_mesh,
-            "distributed_config": distributed_config,
-            "pipeline_config": pipeline_config,
-            "sdpa_method": sdpa_method,
-        }
-
-        if cfg_qat is not None and cfg_qat.get("enabled", False):
-            if cfg_peft is not None:
-                raise ValueError("QAT with PEFT is not currently supported")
-            qat_config_attr = getattr(cfg_qat, "qat_config", None)
-            if qat_config_attr is not None:
-                kwargs["qat_config"] = qat_config_attr.instantiate()
-            else:
-                # Fallback to legacy quantizer format for backward compatibility
-                quantizer_attr = getattr(cfg_qat, "quantizer", None)
-                if quantizer_attr is not None:
-                    kwargs["qat_config"] = quantizer_attr.instantiate()
-
-        if cfg_moe is not None:
-            from nemo_automodel.components.moe.config import MoEParallelizerConfig
-
-            if isinstance(cfg_moe, MoEParallelizerConfig):
-                kwargs["moe_config"] = cfg_moe
-            else:
-                moe_dict = cfg_moe.to_dict() if hasattr(cfg_moe, "to_dict") else dict(cfg_moe)
-                # activation_checkpointing is handled separately; strip config keys
-                moe_dict.pop("activation_checkpointing", None)
-                moe_dict.pop("_target_", None)
-                kwargs["moe_config"] = MoEParallelizerConfig(**moe_dict)
-            kwargs["activation_checkpointing"] = activation_checkpointing
-
-        if cfg_fp8 is not None:
-            kwargs["fp8_config"] = build_fp8_config(cfg_fp8)
-        if cfg_compile is not None:
-            kwargs["compile_config"] = build_compile_config(cfg_compile)
-        if cfg_quantization is not None:
-            logger.info("Model weight quantization enabled with BitsAndBytes")
-            from nemo_automodel.components.quantization.qlora import create_bnb_config
-
-            kwargs["quantization_config"] = create_bnb_config(cfg_quantization)
-
-        is_nemo_auto_model = cfg_model.get("_target_", None) in (
-            NeMoAutoModelForCausalLM.from_config,
-            NeMoAutoModelForCausalLM.from_pretrained,
-            NeMoAutoModelForSequenceClassification.from_config,
-            NeMoAutoModelForSequenceClassification.from_pretrained,
-        )
-
-        if is_nemo_auto_model:
-            # NeMoAutoModel handles infrastructure internally
-            model = cfg_model.instantiate(**kwargs)
-        else:
-            # For non-NemoAutoModel entry points (e.g., build_gpt2_model),
-            # instantiate the model first, then apply infrastructure separately.
-            # Note: sdpa_method is not supported here — SDPA patching only runs
-            # inside NeMoAutoModel._build_model.
-            if sdpa_method is not None:
-                logger.warning("sdpa_method is ignored for non-NeMoAutoModel targets.")
-            # We must convert config objects into runtime objects (model_wrapper,
-            # autopipeline, parallelize_fn, etc.) via instantiate_infrastructure,
-            # exactly as from_pretrained/from_config do internally.
-            model = cfg_model.instantiate()
-
-            mesh = MeshContext.from_meshes(device_mesh, moe_mesh)
-            model_wrapper, autopipeline, parallelize_fn, qat_quantizer = instantiate_infrastructure(
-                distributed_config=distributed_config,
-                pipeline_config=pipeline_config,
-                qat_config=kwargs.get("qat_config"),
-                moe_config=kwargs.get("moe_config"),
-                activation_checkpointing=kwargs.get("activation_checkpointing", False),
-                device=torch.device("cuda", torch.cuda.current_device()),
-                mesh=mesh,
-            )
-            loss_fn = pipeline_config.loss_fn if pipeline_config is not None else None
-
-            model = apply_model_infrastructure(
-                model,
-                is_meta_device=False,
-                device=torch.cuda.current_device(),
-                mesh=mesh,
-                model_wrapper=model_wrapper,
-                autopipeline=autopipeline,
-                parallelize_fn=parallelize_fn,
-                qat_quantizer=qat_quantizer,
-                loss_fn=loss_fn,
-                peft_config=kwargs.get("peft_config"),
-                fp8_config=kwargs.get("fp8_config"),
-                compile_config=kwargs.get("compile_config"),
-                quantization_config=kwargs.get("quantization_config"),
-                pretrained_model_name_or_path=None,
-                load_base_model=False,
-                cache_dir=hf_constants.HF_HUB_CACHE,
-            )
-
-    # Explicitly unfreeze specified modules (e.g. task heads) that need full fine-tuning
-    if unfreeze_modules:
-        for name, param in model.named_parameters():
-            if any(module_name in name for module_name in unfreeze_modules):
-                param.requires_grad_(True)
-        logging.info(f"Unfroze parameters matching: {unfreeze_modules}")
-
-    return model
-
-
-def compute_trust_remote_code_from_model(cfg_model):
-    """Compute the value of trust_remote_code based on the model configuration.
-
-    Args:
-        cfg_model (ConfigNode): Model configuration.
-
-    Returns:
-        bool: Whether to trust remote code.
-    """
-    if hasattr(cfg_model, "trust_remote_code"):
-        return getattr(cfg_model, "trust_remote_code")
-    elif hasattr(cfg_model, "config") and hasattr(cfg_model.config, "trust_remote_code"):
-        return getattr(cfg_model.config, "trust_remote_code")
-    return resolve_trust_remote_code(_get_model_name(cfg_model))
-
-
-def _build_tokenizer(cfg_model, cfg_ds):
-    trust_remote_code = compute_trust_remote_code_from_model(cfg_model)
-    # if tokenizer is not provided, use the model config to instantiate it
-    if "tokenizer" not in cfg_ds and _get_model_name(cfg_model) is not None:
-        logging.info("Using model config to instantiate tokenizer")
-        tokenizer = NeMoAutoTokenizer.from_pretrained(_get_model_name(cfg_model), trust_remote_code=trust_remote_code)
-    elif cfg_ds.get("tokenizer", None) is None:
-        tokenizer = None
-    elif "_target_" not in cfg_ds.tokenizer:
-        tokenizer_dict = cfg_ds.tokenizer.to_dict()
-        trust_remote_code = tokenizer_dict.pop("trust_remote_code", trust_remote_code)
-        tokenizer = NeMoAutoTokenizer.from_pretrained(**tokenizer_dict, trust_remote_code=trust_remote_code)
-    else:
-        trust_remote_code = cfg_ds.tokenizer.to_dict().pop("trust_remote_code", trust_remote_code)
-        tokenizer = cfg_ds.tokenizer.instantiate(trust_remote_code=trust_remote_code)
-
-    # Finally, check if the dataset target accepts a tokenizer parameter
-    kwargs = {}
-    if tokenizer is not None and callable(cfg_ds._target_):
-        try:
-            sig = inspect.signature(cfg_ds._target_)
-            if "tokenizer" in sig.parameters:
-                kwargs["tokenizer"] = tokenizer
-        except (ValueError, TypeError):
-            # If we can't get the signature, skip adding tokenizer
-            pass
-    return kwargs, tokenizer
-
-
-def build_dataloader(
-    cfg_ds,
-    cfg_dl,
-    cfg_model,
-    cfg_ps,
-    seed,
-    local_batch_size,
-    global_batch_size,
-    max_steps,
-    val_check_interval,
-    dp_rank,
-    dp_world_size,
-    pp_enabled,
-    cp_size=1,
-    model: Optional[nn.Module] = None,
-) -> tuple[DataLoader, PreTrainedTokenizerBase]:
-    """Build a DataLoader for the dataset.
-
-    Args:
-        cfg_ds: Dataset configuration.
-        cfg_dl: DataLoader configuration.
-        cfg_model: Model configuration.
-        cfg_ps: Packed sequence configuration.
-        seed: Random seed.
-        local_batch_size: Local batch size.
-        global_batch_size: Global batch size.
-        max_steps: Maximum number of steps.
-        val_check_interval: Validation check interval.
-        dp_rank: Data parallel rank.
-        dp_world_size: Data parallel world size.
-        pp_enabled: Whether pipeline parallelism is enabled.
-        cp_size: Context parallel size.
-        model: Optional model instance. If provided and packed sequences are enabled,
-            seq_lens will only be included if the model's forward() accepts it.
-    Returns:
-        The instantiated DataLoader and tokenizer.
-    """
-    with ScopedRNG(seed=seed, ranked=True):
-        kwargs, tokenizer = _build_tokenizer(cfg_model, cfg_ds)
-        # Megatron specific kwargs
-        if cfg_ds._target_ == MegatronPretraining:
-            kwargs["global_batch_size"] = global_batch_size
-            kwargs["trainer_max_steps"] = max_steps if max_steps is not None else None
-            kwargs["trainer_val_check_interval"] = val_check_interval
-            ds = cfg_ds.instantiate(**kwargs)
-            ds.build()
-        else:
-            with FirstRankPerNode():
-                ds = cfg_ds.instantiate(**kwargs)
-
-        # If using an IterableDataset, per-rank sharding for unique samples
-        if isinstance(ds, IterableDataset):
-            if callable(getattr(ds, "shard", None)):
-                ds = ds.shard(dp_world_size, dp_rank)
-                logging.info(f"Sharded IterableDataset via dataset.shard: world_size={dp_world_size}, rank={dp_rank}")
-            elif hasattr(ds, "dataset"):
-                # HuggingFace streaming datasets: split by file shards when possible.
-                from datasets.distributed import split_dataset_by_node
-
-                assert hasattr(ds, "dataset"), "dataset must have a dataset attribute"
-                ds.dataset = split_dataset_by_node(ds.dataset, world_size=dp_world_size, rank=dp_rank)
-                logging.info(f"Sharded dataset via split_dataset_by_node: world_size={dp_world_size}")
-            else:
-                logging.warning("IterableDataset does not support sharding; Data may be duplicated across ranks.")
-
-        packed_sequence_size = getattr(cfg_ps, "packed_sequence_size", 0)
-        packing_strategy = getattr(cfg_ps, "packing_strategy", "thd")
-
-        # check if packed sequence is supported (only for thd strategy)
-        supports_seq_lens = _supports_seq_lens(model)
-        if packed_sequence_size > 0 and packing_strategy == "thd" and not supports_seq_lens:
-            logging.warning("Packed sequence is not supported without seq_lens; disabling packed sequence")
-            packed_sequence_size = 0
-
-        # Apply packing if configured
-        if packed_sequence_size > 0:
-            logger.info(f"Packing dataset with size: {packed_sequence_size}, strategy: {packing_strategy}")
-            if hasattr(ds, "shuffle"):
-                ds = ds.shuffle(seed)
-
-            if packing_strategy == "neat":
-                from nemo_automodel.components.datasets.llm.neat_packing import neat_pack_dataset
-                from nemo_automodel.components.datasets.utils import neat_packed_collater
-                from nemo_automodel.components.models.common.packing import configure_packing, get_attn_implementation
-
-                ds = neat_pack_dataset(
-                    ds,
-                    split=cfg_ds.split,
-                    pack_size=packed_sequence_size,
-                    max_packs=getattr(cfg_ps, "max_packs", None),
-                    padding_idx=getattr(tokenizer, "pad_token_id", 0),
-                    drop_long_samples=getattr(cfg_ps, "drop_long_samples", True),
-                )
-                _attn_impl = get_attn_implementation(cfg_model)
-                configure_packing(attn_implementation=_attn_impl)
-                # Set collater with attn_implementation so it produces the right mask format
-                cfg_dl.collate_fn = lambda batch, _ai=_attn_impl: neat_packed_collater(batch, attn_implementation=_ai)
-                logger.info(f"Configured neat packing for attn_implementation={_attn_impl}")
-            else:
-                # "thd" — existing packing logic
-                ds = pack_dataset(
-                    ds,
-                    split=cfg_ds.split,
-                    packed_sequence_size=packed_sequence_size,
-                    max_packs=getattr(cfg_ps, "max_packs", None),
-                    padding_idx=getattr(tokenizer, "pad_token_id", 0),
-                    cp_size=cp_size,
-                )
-
-        if isinstance(ds, MegatronPretraining):
-            ds = ds.get_dataset(split=cfg_ds.splits_to_build)
-            dataloader_type = cfg_dl.get("dataloader_type", "single")
-            if "dataloader_type" in cfg_dl:
-                del cfg_dl.dataloader_type
-            batch_sampler = create_megatron_sampler(
-                dataset_len=len(ds),
-                micro_batch_size=local_batch_size,
-                global_batch_size=global_batch_size,
-                dataloader_type=dataloader_type,
-                rank=dp_rank,
-                world_size=dp_world_size,
-            )
-            dl_kwargs = {"batch_sampler": batch_sampler}
-        elif not isinstance(ds, IterableDataset):
-            shuffle = cfg_dl.get("shuffle", True)
-            if "shuffle" in cfg_dl:
-                del cfg_dl.shuffle
-
-            group_by_length = cfg_dl.get("group_by_length", False)
-            if "group_by_length" in cfg_dl:
-                del cfg_dl.group_by_length
-
-            if group_by_length:
-                from nemo_automodel.components.datasets.llm.length_grouped_sampler import (
-                    LengthGroupedSampler as LLMLengthGroupedSampler,
-                )
-
-                sampler = LLMLengthGroupedSampler(
-                    dataset=ds,
-                    batch_size=local_batch_size,
-                    seed=seed,
-                    num_replicas=dp_world_size,
-                    rank=dp_rank,
-                )
-            else:
-                dist_sampler_kwargs = {
-                    "num_replicas": dp_world_size,
-                    "rank": dp_rank,
-                    "shuffle": shuffle,
-                }
-                sampler = StatefulDistributedSampler(
-                    ds,
-                    seed=seed,
-                    drop_last=True,
-                    **dist_sampler_kwargs,
-                )
-            dl_kwargs = {"sampler": sampler, "batch_size": local_batch_size}
-            if pp_enabled:
-                dl_kwargs["drop_last"] = True
-        else:
-            logging.info("Using IterableDataset; skipping sampler.")
-            # Optional shuffle for streaming IterableDataset (uses HF dataset shuffle if available)
-            shuffle = cfg_dl.get("shuffle", False)
-            shuffle_buffer_size = cfg_dl.get("shuffle_buffer_size", 10000)
-            # Do not pass shuffle-related kwargs to the DataLoader when using IterableDataset
-            # But leave them in dl config to be consistent
-            if hasattr(cfg_dl, "shuffle"):
-                del cfg_dl.shuffle
-            if hasattr(cfg_dl, "shuffle_buffer_size"):
-                del cfg_dl.shuffle_buffer_size
-
-            if shuffle and hasattr(ds, "shuffle"):
-                try:
-                    ds = ds.shuffle(buffer_size=shuffle_buffer_size, seed=seed)
-                    logging.info(f"Shuffling IterableDataset with buffer_size={shuffle_buffer_size}, seed={seed}")
-                except Exception as e:
-                    logging.warning(f"IterableDataset shuffle skipped due to error: {e}")
-            dl_kwargs = {}
-
-        # Handle collate_fn with optional mask precomputation for pipeline parallelism
-        dl_kwargs = dl_kwargs | {"dataset": ds}
-
-        # Handle collate_fn instantiation if it's a ConfigNode
-        if hasattr(cfg_dl, "collate_fn"):
-            if hasattr(cfg_dl.collate_fn, "_target_"):
-                collate_cfg = cfg_dl.collate_fn
-                dl_kwargs["collate_fn"] = lambda batch: collate_cfg.instantiate(batch=batch)
-            else:
-                dl_kwargs["collate_fn"] = cfg_dl.collate_fn
-            assert callable(dl_kwargs["collate_fn"]), "collate_fn must be callable"
-
-        # Chain with mask precomputation if PP is enabled
-        if pp_enabled:
-            from nemo_automodel.components.datasets.utils import add_causal_masks_to_batch
-
-            try:
-                hf_model_config = AutoConfig.from_pretrained(
-                    _get_model_name(cfg_model), trust_remote_code=compute_trust_remote_code_from_model(cfg_model)
-                )
-            except Exception:
-                logger.warning(
-                    "Failed to load model config for causal mask precomputation. "
-                    "Pipeline parallel mask precomputation will be skipped."
-                )
-            else:
-                if "collate_fn" in dl_kwargs:
-                    # Case 1: PP enabled + collate_fn exists -> chain them
-                    # base_collate_fn -> add_causal_masks_to_batch
-                    base_collate_fn = dl_kwargs["collate_fn"]
-
-                    def chained_collate_fn(batch, base_fn=base_collate_fn, config=hf_model_config):
-                        batch = base_fn(batch)  # Apply base collate (padding, batching, etc.)
-                        batch = add_causal_masks_to_batch(batch, model_config=config)  # Add masks
-                        return batch
-
-                    dl_kwargs["collate_fn"] = chained_collate_fn
-                else:
-                    # Case 2: PP enabled + no collate_fn -> only add masks
-                    dl_kwargs["collate_fn"] = lambda batch, config=hf_model_config: add_causal_masks_to_batch(
-                        batch, model_config=config
-                    )
-
-        try:
-            import torch.multiprocessing as mp
-
-            if mp.get_start_method(allow_none=True) is None:
-                mp.set_start_method("spawn", force=True)
-        except RuntimeError:
-            pass
-        return cfg_dl.instantiate(**dl_kwargs), tokenizer
-
-
-def calculate_loss(loss_fn, **kwargs) -> torch.Tensor:
-    """Calculate the loss.
-
-    Args:
-        loss_fn: Loss function.
-        **kwargs: Keyword arguments for the loss function.
-
-    Returns:
-        The loss.
-    """
-    loss_fn_kwargs = {"num_label_tokens": kwargs.pop("num_label_tokens", None)}
-    if isinstance(loss_fn, FusedLinearCrossEntropy):
-        model = kwargs.pop("model")
-        labels = kwargs.pop("labels")
-
-        # find the lm_head in the model
-        lm_head = None
-        if hasattr(model, "get_output_embeddings"):
-            lm_head = model.get_output_embeddings().weight
-        else:
-            for n, p in model.named_parameters(remove_duplicate=False):
-                if "lm_head" in n and n.endswith(".weight"):
-                    lm_head = p
-                    break
-        if lm_head is None:
-            raise ValueError("lm_head.weight not found in model")
-
-        # unshard the possibly sharded lm_head
-        lm_head = lm_head.full_tensor() if hasattr(lm_head, "full_tensor") else lm_head
-        loss_fn_kwargs.update(
-            {
-                "hidden_states": kwargs.pop("hidden_states"),
-                "labels": labels,
-                "lm_weight": lm_head,
-            }
-        )
-    else:
-        loss_fn_kwargs.update(
-            {
-                "logits": kwargs.pop("logits"),
-                "labels": kwargs.pop("labels"),
-            }
-        )
-
-    return loss_fn(**loss_fn_kwargs)
-
-
-def calculate_mtp_loss(
-    loss_fn,
-    *,
-    mtp_per_depth_h: list[torch.Tensor],
-    labels: torch.Tensor,
-    model: nn.Module,
-    scaling_factor: float = 0.1,
-    num_label_tokens: Optional[int] = None,
-    ignore_index: int = -100,
-) -> torch.Tensor:
-    """Compute the DeepSeek-V3 Multi-Token Prediction auxiliary loss.
-
-    Each depth's CE is dispatched through :func:`calculate_loss` with the
-    same loss class as the main path, so MTP inherits FusedLinearCrossEntropy
-    / MaskedCrossEntropy memory and numerical characteristics.
-
-    Args:
-        loss_fn: Configured per-token loss class (same instance the main
-            path uses).
-        mtp_per_depth_h: Per-depth hidden states from the model's MTP head,
-            one ``[B, S, H]`` tensor per depth.
-        labels: Original (unshifted) labels.
-        model: The wrapped model; used to fetch the shared LM head when the
-            loss class needs materialized logits (non-FusedLinearCE path).
-        scaling_factor: Coefficient applied to the summed per-depth CE.
-        num_label_tokens: Total non-ignore label tokens (forwarded to the
-            base loss for sum-reduction normalization).
-        ignore_index: Label value masked out of the CE loss for the trailing
-            ``k+1`` rolled positions at depth ``k``.
-
-    Returns:
-        Scalar MTP loss with autograd graph.
-    """
-    from nemo_automodel.components.models.common.mtp import roll_tensor
-
-    D = len(mtp_per_depth_h)
-    cur_labels = labels
-    total = mtp_per_depth_h[0].new_zeros(())
-    for k, h_k in enumerate(mtp_per_depth_h):
-        cur_labels = roll_tensor(cur_labels, shifts=-1, dim=-1)
-        masked = cur_labels.clone()
-        n_invalid = min(k + 1, masked.shape[-1])
-        masked[..., -n_invalid:] = ignore_index
-
-        if isinstance(loss_fn, FusedLinearCrossEntropy):
-            depth_loss = calculate_loss(
-                loss_fn,
-                hidden_states=h_k,
-                labels=masked,
-                model=model,
-                num_label_tokens=num_label_tokens,
-            )
-        else:
-            depth_loss = calculate_loss(
-                loss_fn,
-                logits=model.get_output_embeddings()(h_k),
-                labels=masked,
-                model=model,
-                num_label_tokens=num_label_tokens,
-            )
-        total = total + depth_loss
-
-    return total * (scaling_factor / D)
-
-
-def build_validation_dataloader(cfg, dp_world_size, dp_rank, pp_enabled, model: Optional[nn.Module] = None):
-    """Build validation dataloaders from validation dataset config entries."""
-
-    def _prepare_val_ds_name(val_ds_name):
-        val_ds_name = val_ds_name.replace("validation_dataset", "")
-        if len(val_ds_name) > 1 and val_ds_name[0] in ("_", "-", "."):
-            val_ds_name = val_ds_name[1:]
-        if val_ds_name == "":
-            val_ds_name = "default"
-        return val_ds_name
-
-    # Build validation dataloader if the config provides it
-    val_dataloaders = {}
-    for val_ds_name in filter(lambda x: x.startswith("validation_dataset"), cfg.to_dict().keys()):
-        val_ds_cfg = cfg.get(val_ds_name, None)
-        val_ds_name = _prepare_val_ds_name(val_ds_name)
-        val_dataloaders[val_ds_name] = build_dataloader(
-            val_ds_cfg,
-            cfg.validation_dataloader,
-            cfg.model,
-            cfg_ps=cfg.get("packed_sequence", None)
-            if _uses_te_dot_product_attention(cfg.model) and _uses_thd_collater(cfg.dataloader)
-            else None,
-            seed=cfg.get("seed", 42),
-            local_batch_size=cfg.get("step_scheduler.local_batch_size", 1),
-            global_batch_size=cfg.get("step_scheduler.global_batch_size", 1),
-            max_steps=cfg.get("step_scheduler.max_steps", None),
-            val_check_interval=cfg.get("step_scheduler.val_every_steps", None),
-            dp_rank=dp_rank,
-            dp_world_size=dp_world_size,
-            pp_enabled=pp_enabled,
-            cp_size=cfg.get("distributed.cp_size", 1),
-            model=model,
-        )[0]
-
-    return val_dataloaders
+# ── Stateless helpers moved to components ──────────────────────────────
+# Bodies now live under components/ and _transformers/. Re-imported here so
+# external callers (sibling recipes — kd.py, train_seq_cls.py, dllm —, tests,
+# RL frameworks like verl/NeMo-RL) keep working without an import-path change.
+
+from nemo_automodel._transformers.auto_tokenizer import (  # noqa: E402, F401
+    _build_tokenizer,
+    _get_model_name,
+    compute_trust_remote_code_from_model,
+)
+from nemo_automodel.components.datasets.llm.build import (  # noqa: E402, F401
+    _get_num_thd_chunks,
+    _uses_te_dot_product_attention,
+    _uses_thd_collater,
+    build_dataloader,
+    build_validation_dataloader,
+)
+from nemo_automodel.components.loss.calculate import calculate_loss  # noqa: E402, F401
+from nemo_automodel.components.loss.mtp import calculate_mtp_loss  # noqa: E402, F401
+from nemo_automodel.components.training.build import build_model  # noqa: E402, F401
 
 
 # ---------------------------------------------------------------------------
