+++ nemo_automodel/recipes/vlm/finetune.py
@@ -73,358 +73,20 @@ from nemo_automodel.recipes._component_builders import (
 from nemo_automodel.recipes._dist_setup import setup_distributed
 from nemo_automodel.recipes.base_recipe import BaseRecipe
 
+# Re-exports (back-compat for vlm/kd.py and external test imports).
+# Builders/helpers were moved out of this recipe in the engine refactor.
+from nemo_automodel._transformers.auto_tokenizer import _get_model_name  # noqa: E402, F401
+from nemo_automodel.components.datasets.vlm.build import (  # noqa: E402, F401
+    build_vlm_dataloader as build_dataloader,
+)
+from nemo_automodel.components.loss.calculate import calculate_loss  # noqa: E402, F401
+from nemo_automodel.components.training.build import build_vlm_model as build_model  # noqa: E402, F401
+from nemo_automodel.vlm_engine import chunk_vlm_media as _chunk_vlm_media  # noqa: E402, F401
+
 if TYPE_CHECKING:
     from torch.optim import Optimizer
 
 logger = logging.getLogger(__name__)
-
-# ---------------------------
-#  Stateless helper functions
-# ---------------------------
-
-
-def _get_model_name(cfg_model):
-    if cfg_model.get("pretrained_model_name_or_path", None) is not None:
-        return cfg_model.pretrained_model_name_or_path
-    elif cfg_model.get("config", None) is not None:
-        if isinstance(cfg_model.config, str):
-            return cfg_model.config
-        return cfg_model.config.get("pretrained_model_name_or_path", None)
-    else:
-        return None
-
-
-def build_model(
-    cfg_model,
-    cfg_freeze,
-    cfg_peft,
-    seed,
-    cfg_fp8=None,
-    cfg_compile=None,
-    device_mesh=None,
-    moe_mesh=None,
-    distributed_config=None,
-    pipeline_config=None,
-    cfg_moe=None,
-    activation_checkpointing=False,
-) -> tuple[nn.Module | AutoPipeline, list["Optimizer"]]:  # noqa: F821
-    """Build and initialize a model for VLM.
-
-    Returns:
-        The instantiated model and optimizer.
-    """
-    with ScopedRNG(seed=seed, ranked=True):
-        # Build infrastructure kwargs
-        kwargs = {
-            "peft_config": cfg_peft,
-            "device_mesh": device_mesh,
-            "moe_mesh": moe_mesh,
-            "distributed_config": distributed_config,
-            "pipeline_config": pipeline_config,
-            "freeze_config": cfg_freeze.to_dict() if cfg_freeze is not None else None,
-        }
-
-        if cfg_moe is not None:
-            from nemo_automodel.components.moe.config import MoEParallelizerConfig
-
-            if isinstance(cfg_moe, MoEParallelizerConfig):
-                kwargs["moe_config"] = cfg_moe
-            else:
-                moe_dict = cfg_moe.to_dict() if hasattr(cfg_moe, "to_dict") else dict(cfg_moe)
-                # activation_checkpointing is handled separately; strip config keys
-                moe_dict.pop("activation_checkpointing", None)
-                moe_dict.pop("_target_", None)
-                kwargs["moe_config"] = MoEParallelizerConfig(**moe_dict)
-            kwargs["activation_checkpointing"] = activation_checkpointing
-
-        if cfg_fp8 is not None:
-            fp8_config = build_fp8_config(cfg_fp8)
-            kwargs["fp8_config"] = fp8_config
-        if cfg_compile is not None:
-            kwargs["compile_config"] = build_compile_config(cfg_compile)
-
-        # Check if using NeMoAutoModel
-        is_nemo_auto_model = cfg_model.get("_target_", None) in (
-            NeMoAutoModelForImageTextToText.from_config,
-            NeMoAutoModelForImageTextToText.from_pretrained,
-            NeMoAutoModelForMultimodalLM.from_config,
-            NeMoAutoModelForMultimodalLM.from_pretrained,
-            NeMoAutoModelForCausalLM.from_config,
-            NeMoAutoModelForCausalLM.from_pretrained,
-        )
-
-        if is_nemo_auto_model:
-            # NeMoAutoModel handles infrastructure internally
-            model = cfg_model.instantiate(**kwargs)
-        else:
-            raise ValueError(
-                f"VLM finetuning requires NeMoAutoModelForImageTextToText. "
-                f"Got model target: {cfg_model.get('_target_', None)}"
-            )
-    return model
-
-
-def _chunk_vlm_media(
-    pixel_values: torch.Tensor,
-    image_grid: torch.Tensor,
-    batch_size: int,
-    n_microbatches: int,
-    n_images_per_sample: torch.Tensor | None = None,
-) -> tuple[list[torch.Tensor], list[torch.Tensor]]:
-    """Deprecated: use ``nemo_automodel.vlm_engine.chunk_vlm_media``.
-
-    Re-exported here for backwards compatibility with sibling recipes and
-    external callers. The body lives next to :class:`VLMEngine`.
-    """
-    from nemo_automodel.vlm_engine import chunk_vlm_media
-
-    return chunk_vlm_media(pixel_values, image_grid, batch_size, n_microbatches,
-                            n_images_per_sample=n_images_per_sample)
-
-
-def build_dataloader(
-    cfg_ds,
-    cfg_dl,
-    pretrained_model_name_or_path,
-    cfg_processor,
-    device_mesh,
-    seed,
-    local_batch_size,
-    cfg_model=None,
-    cfg_ps=None,
-    get_rope_index=None,
-) -> tuple[DataLoader, ProcessorMixin]:
-    """Build a DataLoader for the VLM dataset.
-
-    Args:
-        cfg_ds: Dataset configuration.
-        cfg_dl: DataLoader configuration.
-        pretrained_model_name_or_path: Pretrained model name or path for processor loading.
-        cfg_processor: Processor configuration or None.
-        device_mesh: Device mesh for distributed training.
-        seed: Random seed.
-        local_batch_size: Local batch size.
-        cfg_model: Model configuration (used to detect attention backend).
-        cfg_ps: Packed sequence configuration (top-level ``packed_sequence:`` section).
-            When provided, takes precedence over ``dataset.packing``.
-        get_rope_index: Optional ``model.get_rope_index`` callable. When provided,
-            VLM neat packing computes mRoPE 3D position IDs per sample so packed
-            mRoPE-aware models (Qwen2.5-VL, Qwen3-VL, ...) preserve multimodal
-            position semantics across pack boundaries instead of falling back to
-            plain 1D positions.
-
-    Returns:
-        The instantiated DataLoader and processor.
-    """
-    dist_sampler_kwargs = {
-        "shuffle": cfg_dl.get("shuffle", True),
-    }
-    if device_mesh is not None:
-        from nemo_automodel.components.distributed.mesh_utils import get_flat_mesh
-
-        dp_mesh = get_flat_mesh(device_mesh, "dp")
-        dist_sampler_kwargs |= {
-            "num_replicas": dp_mesh.size(),
-            "rank": dp_mesh.get_local_rank(),
-        }
-
-    with ScopedRNG(seed=seed, ranked=True):
-        processor = None
-        processor_kwargs = {}
-
-        with FirstRankPerNode():
-            # Ensure the processor has a _target_ attribute too
-            if (
-                cfg_processor is not None
-                and hasattr(cfg_processor, "instantiate")
-                and hasattr(cfg_processor, "_target_")
-            ):
-                processor = cfg_processor.instantiate()
-            elif cfg_processor is not None:
-                processor_kwargs = cfg_processor.to_dict()
-
-            # If no processor was instantiated, try AutoProcessor
-            if processor is None:
-                try:
-                    processor = AutoProcessor.from_pretrained(pretrained_model_name_or_path, **processor_kwargs)
-                except Exception as e:
-                    # AutoProcessor.from_pretrained internally loads AutoConfig. Configs
-                    # whose layer_types length differs from num_hidden_layers trip
-                    # validate_layer_type. The processor itself doesn't depend on
-                    # layer_types, so relax the validator and retry once before giving up.
-                    err = str(e)
-                    if "num_hidden_layers" in err and ("layer_types" in err or "layer types" in err):
-                        from nemo_automodel._transformers.v4_patches.layer_types import (
-                            relax_layer_types_validator,
-                        )
-
-                        relax_layer_types_validator()
-                        try:
-                            processor = AutoProcessor.from_pretrained(pretrained_model_name_or_path, **processor_kwargs)
-                        except Exception as retry_exc:
-                            processor = None
-                            logging.warning(
-                                f"AutoProcessor not available for {pretrained_model_name_or_path} ({retry_exc}). "
-                            )
-                    else:
-                        # Some models do not provide an AutoProcessor
-                        processor = None
-                        logging.warning(f"AutoProcessor not available for {pretrained_model_name_or_path} ({e}). ")
-
-            chat_template_raw = cfg_ds.__dict__.pop("chat_template", None)
-            # Update chat_template if chat_template is given
-            if chat_template_raw is not None and processor is not None:
-                processor.chat_template = _resolve_chat_template(chat_template_raw)
-                processor.tokenizer.chat_template = processor.chat_template
-
-            _path_or_ds = getattr(cfg_ds, "path_or_dataset", None) or cfg_ds.get("path_or_dataset", None)
-            if _path_or_ds is not None:
-                ds = cfg_ds.instantiate(path_or_dataset=_path_or_ds)
-            else:
-                ds = cfg_ds.instantiate()
-
-        # Resolve packing config: top-level packed_sequence (LLM-style) takes
-        # precedence over legacy dataset.packing (backward compat).
-        if cfg_ps is not None:
-            _ps_enabled = getattr(cfg_ps, "pack_size", 0) > 0
-            packing_cfg = cfg_ps if _ps_enabled else None
-            pretokenize = getattr(cfg_ps, "pretokenize", _ps_enabled)
-            max_length = getattr(cfg_ps, "max_length", None)
-        else:
-            _legacy = cfg_ds.get("packing", None)
-            _ps_enabled = _legacy is not None and _legacy.get("enabled", False)
-            packing_cfg = _legacy if _ps_enabled else None
-            max_length = cfg_ds.get("max_length", None)
-            pretokenize = cfg_ds.get("pretokenize", max_length is not None)
-
-        if pretokenize:
-            from nemo_automodel.components.datasets.vlm.collate_fns import pad_collate_fn
-            from nemo_automodel.components.datasets.vlm.datasets import PreTokenizedDatasetWrapper
-
-            ds_raw = ds
-            truncate = cfg_ds.get("truncate", max_length is not None)
-
-            post_tokenize_hook = cfg_ps.get("post_tokenize_hook_fn", None) if cfg_ps is not None else None
-
-            ds = PreTokenizedDatasetWrapper(
-                ds_raw,
-                processor,
-                max_length=max_length,
-                truncate=truncate,
-                post_tokenize_hook=post_tokenize_hook,
-            )
-
-            if packing_cfg:
-                from nemo_automodel.components.datasets.vlm.collate_fns import neat_packed_vlm_collater
-                from nemo_automodel.components.datasets.vlm.neat_packing_vlm import neat_pack_dataset_vlm
-                from nemo_automodel.components.models.common.packing import configure_packing, get_attn_implementation
-
-                ds = neat_pack_dataset_vlm(
-                    ds,
-                    pack_size=packing_cfg.get("pack_size", max_length),
-                    padding_idx=getattr(processor.tokenizer, "pad_token_id", 0) or 0,
-                    drop_long_samples=packing_cfg.get("drop_long_samples", True),
-                    max_packs=packing_cfg.get("max_packs", None),
-                    ds_raw=ds_raw,
-                    packing_ratio=packing_cfg.get("packing_ratio", 1.0),
-                    processor=processor,
-                    balance_media_tokens=packing_cfg.get("balance_media_tokens", True),
-                    get_rope_index=get_rope_index,
-                )
-                _pad_id = getattr(processor.tokenizer, "pad_token_id", 0) or 0
-                _collate_max_length = packing_cfg.get("collate_max_length", None)
-                _attn_impl = get_attn_implementation(cfg_model)
-
-                configure_packing(attn_implementation=_attn_impl)
-                logging.info(f"Configured VLM neat packing for attn_implementation={_attn_impl}")
-
-                collate_fn = lambda examples, _pi=_pad_id, _ml=_collate_max_length, _ai=_attn_impl: (
-                    neat_packed_vlm_collater(
-                        examples,
-                        padding_idx=_pi,
-                        max_length=_ml,
-                        attn_implementation=_ai,
-                    )
-                )
-            else:
-                collate_cfg = cfg_dl.get("collate_fn", None)
-                if collate_cfg:
-                    collate_fn = lambda examples: collate_cfg.instantiate(examples=examples, processor=processor)
-                else:
-                    collate_fn = lambda examples: pad_collate_fn(examples, processor)
-
-            sampler = torch.utils.data.distributed.DistributedSampler(
-                ds,
-                **dist_sampler_kwargs,
-            )
-        else:
-            sampler = torch.utils.data.distributed.DistributedSampler(
-                ds,
-                **dist_sampler_kwargs,
-            )
-            collate_cfg = cfg_dl.get("collate_fn", None)
-            if collate_cfg:
-                collate_fn = lambda examples: collate_cfg.instantiate(examples=examples, processor=processor)
-            else:
-                processor_type = type(processor).__name__
-                if processor_type not in COLLATE_FNS:
-                    logging.warning(f"You are using {processor_type} with default collate function.")
-                    processor_type = "default"
-                collate_fn = lambda examples: COLLATE_FNS[processor_type](examples, processor)
-
-        if hasattr(ds, "robust_collate"):
-            collate_fn = ds.robust_collate(collate_fn)
-
-        return cfg_dl.instantiate(
-            dataset=ds, sampler=sampler, collate_fn=collate_fn, batch_size=local_batch_size
-        ), processor
-
-
-def calculate_loss(loss_fn, **kwargs) -> torch.Tensor:
-    """Calculate the loss.
-
-    Args:
-        loss_fn: Loss function.
-        **kwargs: Keyword arguments for the loss function.
-
-    Returns:
-        The loss.
-    """
-    loss_fn_kwargs = {"num_label_tokens": kwargs.pop("num_label_tokens", None)}
-    if isinstance(loss_fn, FusedLinearCrossEntropy):
-        model = kwargs.pop("model")
-        labels = kwargs.pop("labels")
-
-        # find the lm_head in the model
-        lm_head = None
-        if hasattr(model, "get_output_embeddings"):
-            lm_head = model.get_output_embeddings().weight
-        else:
-            for n, p in model.named_parameters(remove_duplicate=False):
-                if "lm_head" in n and n.endswith(".weight"):
-                    lm_head = p
-                    break
-        if lm_head is None:
-            raise ValueError("lm_head.weight not found in model")
-
-        # unshard the possibly sharded lm_head
-        lm_head = lm_head.full_tensor() if hasattr(lm_head, "full_tensor") else lm_head
-        loss_fn_kwargs.update(
-            {
-                "hidden_states": kwargs.pop("hidden_states"),
-                "labels": labels,
-                "lm_weight": lm_head,
-            }
-        )
-    else:
-        loss_fn_kwargs.update(
-            {
-                "logits": kwargs.pop("logits"),
-                "labels": kwargs.pop("labels"),
-            }
-        )
-
-    return loss_fn(**loss_fn_kwargs)
 
 
 # ---------------------------------------------------------------------------
